Updated - Sprites needs fixing
All the attack sprites are way too small and when scaled, they will be scaled weirdly when it comes to animating as they will scale the gameobject itself instead of the sprite separately.
Otherwise, I have the game updated to have the zap shield as explosion only at one by one with a cd of 10 secs.
Z and X are now both basic attacks with 1 dmg each slash.